Shut up most warnings, mostly about signedness of data types
git-svn-id: svn://svn.savannah.gnu.org/mtools/trunk@741 8fbfe7c3-7cdf-472d-ae67-d119146b2a64

+++ dirCache.c
@@ -46,7 +46,7 @@ static uint32_t calcHash(wchar_t *name)
 				     * successive letters cannot cover each
 				     * other easily */
 		c = towupper((wint_t)*name);
-		hash ^=  (c * (c+2)) ^ (i * (i+2));
+		hash ^= (uint32_t) (c * (c+2)) ^ (i * (i+2));
 		hash &= 0xffffffff;
 		i++;
 		name++;
